+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Version 0.0.25 (2024-08-24)
+-------------
+
+- Added Comment model and integrated Pydantic schemas for comment
+  management, including validation, pagination,and sorting.
+- Added full set of CRUD operations for comments, encompassing creation,
+  retrieval, updating, and deletion functionalities.
+
 Version 0.0.24 (2024-08-24)
 -------------